+++ src/components/Search.vue
@@ -73,7 +73,7 @@
 <script>
 import axios from 'axios'
 import List from './Search/List'
-import {API_URL} from './../util/App'
+import {API_URL} from './../constant/App'
 
 export default {
   name: 'Search',
